Refactor the logic to show and hide additional steps into service§
We need to go further than only looking to the referenced step in order to show or hide it. We need to also check that that step doesn't include any other references. Before we do this more complex tree traversal operation we refactor the logic out of the model and into a service.
