+++ app/services/save_answer.rb
@@ -17,8 +17,7 @@ class SaveAnswer
 
     if answer.valid?
       answer.save
-      answer.step.check_to_show_additional_step!
-      answer.step.check_to_hide_additional_step!
+      ToggleAdditionalSteps.new(step: answer.step).call
       result.success = true
     end
